Port the dcrd 2.2 rpcserver delta and wire the live ban flow
- addnode/node map connection failures through dcrd 2.2's branches:
  the synchronous dial has no context, so every dial error takes the
  internal-error branch with the raw message (the prefix is
  log-only); node's disconnect/remove keep their per-arm
  invalid-parameter fallthroughs
- gettreasuryspendvotes carries the u32 vote tallies through the
  TSpendCountVotes seam, and the live chain adapter now implements it
  (previously an unimplemented panic); the ADXR logging subsystem is
  gone per master
- the misbehavior handlers feed a shared banned-host map with
  BanPeer's semantics at every dcrd ban site: threshold-tripped
  mempool/getdata/notfound scores (notfound intake newly dispatched
  and forwarded to the sync manager), the direct committed-filter ban
  replacing the old +100 score, and the direct empty/invalid message
  bans keeping dcrd's only-disconnect-inside-BanPeer behavior;
  whitelisted scores accumulate per 2.2 without ever banning
- both inbound and outbound admission refuse banned hosts before any
  handshake work, clearing expired bans as consulted
- dcrd 2.2 retires the legacy mining-state exchange: getminings and
  miningstate ban at protocol versions covering getinitstate, repeats
  of either state request ban behind the sent latches, and repeated
  initial state messages ban behind a new received latch, with the
  initstate intake forwarded like OnInitState
- transport read errors are typed so the read loop mirrors OnRead's
  wire-protocol-violation ban; serve signals carry owned reasons
- pinned by end-to-end tests driving a misbehavior ban and a
  malformed frame through real connections (recorded host, refused
  reconnect), rewritten srvabuse/srvinitstate/srvrelay rows, and a
  native node-connect error case; the addnode dial-failure and
  debuglevel help rows updated in place

+++ crates/dcroxide-rpc/src/server.rs
@@ -300,12 +300,13 @@ pub trait RpcChain {
         unimplemented!("lottery_data_for_block")
     }
     /// The (yes, no) vote tally for the treasury spend counted up to
-    /// the given block (dcrd `TSpendCountVotes`).
+    /// the given block (dcrd `TSpendCountVotes`; u32 tallies per
+    /// dcrd 2.2).
     fn tspend_count_votes(
         &mut self,
         _check_block: &Hash,
         _tspend: &MsgTx,
-    ) -> Result<(i64, i64), TSpendCountVotesFailure> {
+    ) -> Result<(u32, u32), TSpendCountVotesFailure> {
         unimplemented!("tspend_count_votes")
     }
 }
